Login Fixed - Keyboard Update
—Login Fixed—
+ Delete User now deletes until a) you run out of users or b) you click
“Cancel Deletion”
+ Create User has now been properly updated - Before the process was
deleting the user, saving, clearing all of the lists, and then
re-loading every single user. The new process saves the index in which
they were at and recursively subtracts 5 from their current index
unless they are at or below 5. This way we can save the layers they
were at and cycle through the rest of the users untouched by the
deletion. It is a lot faster and a lot easier to follow up on in case
there is another bug.
+ Implemented default list for the keyboard from the inspector if the
user has no prior keyboard data saved.

- - This will be added in future updates (Maybe) in which you will be
able to click a delete button that just deletes all of the data in the
folder that the keyboard and menu data are holding.

—Next Update—
- Character movement
- AI (Start development of paths and sight/ hearing)
- Clean up options menu before overhaul